requirements.txt
flask
beautifulsoup4
requests
googlesearch-python